Commit temporarily, datacenter does half
