+++ app/main/vcenter/db/__init__.py
@@ -7,3 +7,6 @@ from app.main.vcenter.db import disks
 from app.main.vcenter.db import images
 from app.main.vcenter.db import snapshots
 from app.main.vcenter.db import resource_pool
+from app.main.vcenter.db import datacenters
+from app.main.vcenter.db import resource_pool
+
